Merge and update AskUbuntu module

+++ Modules/AskUbuntuModuleTest.py
@@ -12,21 +12,18 @@ def main(argv):
     datafile = sys.argv[1]
 
     fileDir = os.path.dirname(os.path.realpath('__file__'))
-    # Check for the Q/A dict stored off
-    qfilename = os.path.join(fileDir, '../Data/ubuntuQDict.txt.gz')
-    afilename = os.path.join(fileDir, '../Data/ubuntuADict.txt.gz')
-    dfilename = os.path.join(fileDir, datafile)
+    qfilename = os.path.join(fileDir, '../Data/ubuntuQDict.txt.gz');
+    qidffilename = os.path.join(fileDir, '../Data/ubuntuQIdfDict.txt.gz');
+    afilename = os.path.join(fileDir, '../Data/ubuntuADict.txt.gz');
+    dfilename = os.path.join(fileDir, datafile);
 
-    if (os.path.exists(qfilename) == True and
-       os.path.exists(afilename) == True):
-        print("Question/Answer Dicts found! Loading...")
+    # Check for the saved dictionaries
+    if (os.path.exists(qfilename) == True):
+        print("Question Dict found! Loading...")
         # Load our Question dictionary
-        QDict = pickle.load(gzip.open(qfilename,'rb'))
-        #QDict = pickle.load(open(qfilename, "rb"));
-        # Load our Answer dictionary
-        ADict = pickle.load(gzip.open(afilename, 'rb'))
+        QDict = pickle.load(gzip.open(qfilename,'rb'));
     else:
-        print("Question/Answer Dicts NOT found! Creating...")
+        print("Question Dict NOT found! Creating...")
         #Open the dataFile
         data = ET.parse(dfilename)
         #Get the questions
@@ -35,11 +32,32 @@ def main(argv):
         with gzip.open(qfilename, 'wb') as f:
             f.write(pickle.dumps(QDict))
 
+    if (os.path.exists(afilename) == True):
+        print("Answer Dict found! Loading...")
+        # Load our Answer dictionary
+        ADict = pickle.load(gzip.open(afilename, 'rb'));
+    else:
+        print("Answer Dict NOT found! Creating...")
+        #Open the dataFile
+        data = ET.parse(dfilename);
+        #Get the questions
+        QDict, ansArr = StringMatch.loadQStrs(data);
         #Get the answers
         ADict = StringMatch.loadAStrs(data, ansArr)
         # Save the answers
         with gzip.open(afilename, 'wb') as f:
                 f.write(pickle.dumps(ADict))
+
+    if (os.path.exists(qidffilename) == True):
+        print("Question IDF Dict found! Loading...")
+        # Load our Question IDF dictionary
+        QIdfDict = pickle.load(gzip.open(qidffilename,'rb'));
+    else:
+        print("Question IDF Dict NOT found! Creating...")
+        QIdfDict = IdfMatch.genQIdfData(QDict);
+        # Save the Question IDF
+        with gzip.open(qidffilename, 'wb') as f:
+                f.write(pickle.dumps(QIdfDict))
 
     #Ask user for a question
     inStr = input("Ask an Ubuntu Question: ");
@@ -48,17 +66,19 @@ def main(argv):
     #Should get index, matched question, and ratio values back
     # wordMatch
     ##wordRatio, wordQ, wordIdx = StringMatch.wordMatch(inStr, QDict);
-    QIdfDict = IdfMatch.genQIdfData(QDict);
-    wordRatio, wordQ, wordIdx = IdfMatch.wordMatchIdf(inStr, QDict, QIdfDict);
-
-    #Get answer based on index
-    wordAns = ADict[wordIdx];
+    # wordIdfMatch
+    wordRatios, wordQs, wordIdxs = IdfMatch.wordMatchIdf(inStr, QDict, QIdfDict);
 
     #Print everything out
     print("\nQuestion Asked: ",inStr);
-    print("Ratio: ", wordRatio*100);
-    print("Closest Question Found (Word): ",wordQ);
-    print("Answer: ", wordAns);
+    wordAns = {};
+    #Get answer based on index
+    for i in range(3):
+      wordAns[i] = ADict[wordIdxs[i]];
+      print("Showing Match Number: ", i+1);
+      print("\nRatio: ", wordRatios[i]*100);
+      print("\nQuestion Found: ",wordQs[i]);
+      print("\nAnswer: ", wordAns[i]);
 
 if __name__ == "__main__":
     #Argument one should be relative path to datafile (../Data/Posts.xml)
